fix: Converting the two sides of the website into two separate containers
